+++ HISTORY.md
@@ -3,13 +3,17 @@ v0.21 (xxx/xx)
 
 Additions:
 
+- dictators:
+    - `--dictator=<key>`
+    - can always post and vote, and always wins forks
 - open chains:
-    - no `--pioneer`
+    - no `--pioneer`, `--dictator`
     - anyone may post and vote, including unsigned
 
 Modifications:
 
 - chain alias: `/chat` (was `'#chat'`)
+- genesis: `dictators:` and `pioneers:` sections
 
 v0.20 (aug/26)
 --------------